# docs/swagger.yaml
basePath: /
definitions:
  internal_adapters_primary_api_address.addAddressReq:
    properties:
      city:
        type: string
      country:
        type: string
      line1:
        type: string
      postal_code:
        type: string
      province:
        type: string
    type: object
  internal_adapters_primary_api_address.addAddressResp:
    properties:
      id:
        type: string
    type: object
  internal_adapters_primary_api_address.addressInfoResp:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: string
      is_default:
        type: boolean
      line1:
        type: string
      postal_code:
        type: string
      province:
        type: string
    type: object
  internal_adapters_primary_api_address.getDefaultAddressResp:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: string
      line1:
        type: string
      postal_code:
        type: string
      province:
        type: string
    type: object
  internal_adapters_primary_api_address.listAddressesResp:
    properties:
      addresses:
        items:
          $ref: '#/definitions/internal_adapters_primary_api_address.addressInfoResp'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    A RESTful e-commerce API built with Go using hexagonal architecture.
    Features: User authentication (JWT), Product management, Order processing, Address management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
  /addresses:
    get:
      consumes:
      - application/json
      description: Get all addresses for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_address.listAddressesResp'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all addresses
      tags:
      - Addresses
    post:
      consumes:
      - application/json
      description: Create a new address for the authenticated user
      parameters:
      - description: Address data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_address.addAddressReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_address.addAddressResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add a new address
      tags:
      - Addresses
  /addresses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an address by ID
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete an address
      tags:
      - Addresses
  /addresses/{id}/default:
    put:
      consumes:
      - application/json
      description: Set an address as the default address
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Set default address
      tags:
      - Addresses
  /addresses/default:
    get:
      consumes:
      - application/json
      description: Get the default address for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_address.getDefaultAddressResp'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get default address
      tags:
      - Addresses
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"